refactor: add constants in a separate file
