+++ src/utils/constants.ts
@@ -1,0 +1,18 @@
+export const CODES = {
+  OK: 200,
+  CREATE: 201,
+  DELETE: 204,
+  NOT_FOUND: 404,
+  INVALID: 400,
+  SERVER_ERROR: 500,
+};
+
+export const MESSAGES = {
+  INVALID_UUID: 'Invalid UUID',
+  USER_NOT_FOUND: 'User not found',
+  USER_DELETED: 'User successfully deleted',
+  END_POINT_NOT_FOUND: 'Endpoint not found',
+  INVALID_FIELDS: 'Missing or invalid fields',
+  SERVER_ERROR: 'Server error',
+  SERVER_RUNNING: 'Server running on',
+};
