+++ src/utils/helpers.ts
@@ -1,23 +1,5 @@
 import { IncomingMessage, ServerResponse } from 'node:http';
-
-export const CODES = {
-  OK: 200,
-  CREATE: 201,
-  DELETE: 204,
-  NOT_FOUND: 404,
-  INVALID: 400,
-  SERVER_ERROR: 500,
-};
-
-export const MESSAGES = {
-  INVALID_UUID: 'Invalid UUID',
-  USER_NOT_FOUND: 'User not found',
-  USER_DELETED: 'User successfully deleted',
-  END_POINT_NOT_FOUND: 'Endpoint not found',
-  INVALID_FIELDS: 'Missing or invalid fields',
-  SERVER_ERROR: 'Server error',
-  SERVER_RUNNING: 'Server running on',
-};
+import { CODES, MESSAGES } from './constants';
 
 export function setHeader(res: ServerResponse, statusCode: (typeof CODES)[keyof typeof CODES]) {
   res.statusCode = statusCode;
